pyproject.toml
[tool.poetry]
name = "hundoLeaderboard"
version = "0.1.0"
description = "A front end to display the top records for the Recipies At Home project for the TTYD 100% TAS"
authors = ["Lego <lego6245@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
azure-cosmosdb-table = "^1.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
